refactor(checkers): generalize Javascript data flow analysis
- Remove SQL-specific checks to enable broader analysis

BREAKING CHANGE:
- sql_injection_dfg.go renamed to js_dataflow.go
- sql_injection_dfg_test.go renamed to js_dataflow_test.go
